refactor: Restructure cmdparser.cpp to fix PIMPL and namespace errors
This commit comprehensively refactors Manager/src/cmdparser.cpp
to resolve critical build errors related to the PIMPL idiom,
nested namespaces, and other type mismatches.

Key changes:
- Relocated the complete definition of the `CmdParserPrivate` class
  (including all its member function definitions and associated helper
  structs `CmdParserFlags`, `CmdParserCommand`) to appear at the
  beginning of the `AkVCam` namespace scope within `cmdparser.cpp`.
  This ensures `CmdParserPrivate` is a complete type before its use,
  fixing "incomplete type" errors during `new CmdParserPrivate()` and
  when binding methods.
- Removed a redundant, nested `namespace AkVCam` block. All AkVCam-
  specific definitions now reside under a single, top-level `AkVCam`
  namespace, resolving "no member named X in namespace AkVCam::AkVCam"
  errors.
- Corrected an `operator*` usage in `CmdParserPrivate::drawTableHLine`
  by adding `static_cast<size_t>` to ensure the correct overload for
  string repetition is called, fixing "invalid operands to binary
  expression ('std::string' and 'unsigned int/long long')" errors.
- Changed the `VideoFrame` validity check in the splitting thread's lambda
  (within `CmdParserPrivate::startSplittingForPhysicalCamera`) from
  `.isValid()` to direct boolean conversion (`if (m_latestPhysicalFrames[...])`),
  relying on `VideoFrame::operator bool()`, to fix "no member named
  'isValid' in 'AkVCam::VideoFrame'" errors.
- Ensured standard library headers like `<atomic>`, `<mutex>`, and project
  header "icapturedevice.h" are included at the top-level file scope.

This refactoring aims to create a correct and clean structure for
`cmdparser.cpp`, resolving the blocking compilation errors.

+++ Manager/src/cmdparser.cpp
@@ -259,7 +259,7 @@ namespace AkVCam {
         if (toStdErr) out_stream = &std::cerr;
         *out_stream << '+';
         for (const auto &len : columnsLength)
-            *out_stream << std::string("-") * (len + 2) << '+';
+            *out_stream << std::string("-") * static_cast<size_t>(len + 2) << '+'; // Applied static_cast
         *out_stream << std::endl;
     }
